Add SuccessfulOutputsExecutionRecorderTest
